Reorganize configuration
