+++ core/src/common/index.ts
@@ -1,3 +1,2 @@
 export * from './types';
-export * from './config';
 export * from './logger';
